+++ .pre-commit-config.yaml
@@ -19,7 +19,7 @@ repos:
           - "fastapi"
           - "types-requests"
           
-  - repo: https://github.com/Daco2020/auto-commit-msg.git
-    rev: v0.1.3
-    hooks:
-      - id: "auto-commit-msg"
+  # - repo: https://github.com/Daco2020/auto-commit-msg.git
+  #   rev: v0.1.3
+  #   hooks:
+  #     - id: "auto-commit-msg"
